requirements.txt
depthai
ultralytics
opencv-python
requests
pyserial
pynmea2
